+++ linkerd/http/upgrade/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "linkerd-http-upgrade"
+version = "0.1.0"
+authors = ["Linkerd Developers <cncf-linkerd-dev@lists.cncf.io>"]
+license = "Apache-2.0"
+edition = "2021"
+publish = false
+description = """
+Facilities for HTTP/1 upgrades.
+"""
+
+[dependencies]
+bytes = "1"
+drain = "0.1"
+futures = { version = "0.3", default-features = false }
+http = "0.2"
+http-body = "0.4"
+hyper = { version = "0.14", default-features = false, features = ["client"] }
+pin-project = "1"
+tokio = { version = "1", default-features = false }
+tower = { version = "0.4", default-features = false }
+tracing = "0.1"
+try-lock = "0.2"
+
+linkerd-duplex = { path = "../../duplex" }
+linkerd-error = { path = "../../error" }
+linkerd-http-version = { path = "../version" }
+linkerd-io = { path = "../../io" }
+linkerd-stack = { path = "../../stack" }
