+++ linkerd/proxy/http/Cargo.toml
@@ -49,6 +49,7 @@ linkerd-http-insert = { path = "../../http/insert" }
 linkerd-http-override-authority = { path = "../../http/override-authority" }
 linkerd-http-retain = { path = "../../http/retain" }
 linkerd-http-stream-timeouts = { path = "../../http/stream-timeouts" }
+linkerd-http-upgrade = { path = "../../http/upgrade" }
 linkerd-http-version = { path = "../../http/version" }
 linkerd-io = { path = "../../io" }
 linkerd-proxy-balance = { path = "../balance" }
